modernize OpenClaw integration
