+++ openclaw.plugin.json
@@ -1,6 +1,16 @@
 {
   "id": "groupme",
   "channels": ["groupme"],
+  "channelEnvVars": {
+    "groupme": [
+      "GROUPME_BOT_ID",
+      "GROUPME_ACCESS_TOKEN",
+      "GROUPME_BOT_NAME",
+      "GROUPME_GROUP_ID",
+      "GROUPME_PUBLIC_DOMAIN",
+      "GROUPME_CALLBACK_URL"
+    ]
+  },
   "configSchema": {
     "type": "object",
     "additionalProperties": false,
