+++ runtime-api.ts
@@ -1,0 +1,3 @@
+export type { PluginRuntime } from "openclaw/plugin-sdk/core";
+export type { ResolvedGroupMeAccount } from "./src/types.js";
+export { getGroupMeRuntime, setGroupMeRuntime } from "./src/runtime.js";
