+++ setup-entry.ts
@@ -1,0 +1,4 @@
+import { defineSetupPluginEntry } from "openclaw/plugin-sdk/core";
+import { groupmeSetupPlugin } from "./api.js";
+
+export default defineSetupPluginEntry(groupmeSetupPlugin);
