+++ src/config-schema.ts
@@ -1,7 +1,7 @@
 import {
   BlockStreamingCoalesceSchema,
   MarkdownConfigSchema,
-} from "openclaw/plugin-sdk";
+} from "openclaw/plugin-sdk/channel-config-schema";
 import { z } from "zod";
 
 const allowFromEntry = z.union([z.string(), z.number()]);
