+++ tests/accounts.test.ts
@@ -1,4 +1,4 @@
-import { DEFAULT_ACCOUNT_ID } from "openclaw/plugin-sdk";
+import { DEFAULT_ACCOUNT_ID } from "openclaw/plugin-sdk/account-id";
 import { afterEach, describe, expect, it } from "vitest";
 import { resolveGroupMeAccount } from "../src/accounts.js";
 import type { CoreConfig } from "../src/types.js";
